requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.0
gunicorn==23.0.0
uvloop==0.21.0
httptools==0.7.1

# Async & Web
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.5.2

# YouTube Integration
youtube-transcript-api==0.6.1
yt-dlp==2025.1.26

# Database & ORM
asyncpg==0.30.0
SQLAlchemy[asyncio]==2.0.35
greenlet==3.3.0
alembic==1.13.3

# Caching & Message Queue
redis[hiredis]==5.0.7
rq==1.15.1
rq-scheduler==0.13.1

# Observability
structlog==24.4.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.21.0
cachetools==5.5.0

# Security
python-dotenv==1.0.1
cryptography==42.0.0
PyJWT==2.10.1
python-multipart==0.0.9

# Error Tracking (Optional)
sentry-sdk[fastapi]==2.20.0

# Utilities
python-dateutil==2.8.2
pytz==2025.1
tenacity==9.0.0
